+++ frontend/tsconfig.json
@@ -35,7 +35,7 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
     "**/*.mts"
-, "src/types/hi"  ],
+  ],
   "exclude": [
     "node_modules"
   ]
